Moving logout functionality to base page
