+++ capacitor.config.json
@@ -10,7 +10,4 @@
     }
   },
   "cordova": {},
-  "server": {
-    "url": "http://192.168.1.111:8100"
-  }
 }
