+++ src/app/home/home.page.html
@@ -1,6 +1,6 @@
 <ion-header>
   <ion-toolbar color="dark">
-    <ion-button color="dark" (click)="logout()">
+    <ion-button color="dark" (click)="onLogout()">
       Logout
     </ion-button>
     <ion-title>DES Ionic 5 - Main Screen</ion-title>
@@ -18,6 +18,5 @@
       Login
     </ion-button>
   </div>
-
 
 </ion-content>
